+++ ffupdater/src/main/java/de/marmaro/krt/ffupdater/background/BackgroundJob.kt
@@ -1,8 +1,6 @@
 package de.marmaro.krt.ffupdater.background
 
-import android.app.DownloadManager
 import android.content.Context
-import android.content.Context.DOWNLOAD_SERVICE
 import android.util.Log
 import androidx.annotation.MainThread
 import androidx.work.*
@@ -73,15 +71,14 @@ class BackgroundJob(context: Context, workerParams: WorkerParameters) :
     @MainThread
     private suspend fun executeBackgroundJob(): Result {
         Log.i(LOG_TAG, "Execute background job for update check.")
-        val downloadManager = applicationContext.getSystemService(DOWNLOAD_SERVICE) as DownloadManager
-        if (DownloadManagerUtil.isDownloadingFilesNow(downloadManager)) {
+        if (DownloadManagerUtil.isDownloadingFilesNow(applicationContext)) {
             Log.i(LOG_TAG, "Retry background job because other downloads are running.")
             return Result.retry()
         }
 
         val appsWithUpdates = findAppsWithUpdates()
         if (NetworkUtil.isActiveNetworkUnmetered(applicationContext) && StorageUtil.isEnoughStorageAvailable()) {
-            downloadUpdatesInBackground(downloadManager, appsWithUpdates)
+            downloadUpdatesInBackground(appsWithUpdates)
         }
         showUpdateNotification(appsWithUpdates)
         updateLastBackgroundCheckTimestamp()
@@ -107,11 +104,8 @@ class BackgroundJob(context: Context, workerParams: WorkerParameters) :
      * If the current network is unmetered, then download the update for the given apps
      * with the DownloadManager in the background.
      */
-    private suspend fun downloadUpdatesInBackground(
-        downloadManager: DownloadManager,
-        appsWithUpdates: List<App>
-    ) {
-        appsWithUpdates.forEach { downloadUpdateInBackground(it, downloadManager) }
+    private suspend fun downloadUpdatesInBackground(appsWithUpdates: List<App>) {
+        appsWithUpdates.forEach { downloadUpdateInBackground(it) }
     }
 
     /**
@@ -119,7 +113,7 @@ class BackgroundJob(context: Context, workerParams: WorkerParameters) :
      * download is finished.
      */
     @MainThread
-    private suspend fun downloadUpdateInBackground(app: App, downloadManager: DownloadManager) {
+    private suspend fun downloadUpdateInBackground(app: App) {
         val appCache = AppCache(app)
         val cachedUpdateChecker = app.detail.updateCheck(applicationContext)
         val availableResult = cachedUpdateChecker.availableResult
@@ -130,14 +124,13 @@ class BackgroundJob(context: Context, workerParams: WorkerParameters) :
 
         Log.i(LOG_TAG, "Download $app in the background.")
         val downloadId = DownloadManagerUtil.enqueue(
-            downloadManager,
             applicationContext,
             app,
             availableResult,
             appCache.getFileName()
         )
         repeat(60 * 60) {
-            val status = DownloadManagerUtil.getStatusAndProgress(downloadManager, downloadId).status
+            val status = DownloadManagerUtil.getStatusAndProgress(applicationContext, downloadId).status
             if (status == SUCCESSFUL) {
                 return
             }
@@ -146,7 +139,7 @@ class BackgroundJob(context: Context, workerParams: WorkerParameters) :
             }
             delay(1000)
         }
-        downloadManager.remove(downloadId)
+        DownloadManagerUtil.remove(applicationContext, downloadId)
         appCache.delete(applicationContext)
     }
 
@@ -161,6 +154,7 @@ class BackgroundJob(context: Context, workerParams: WorkerParameters) :
     companion object {
         private const val WORK_MANAGER_KEY = "update_checker"
         private const val LOG_TAG = "BackgroundJob"
+
         // waiting time = 0.5m + 1m + 2m + 4m + 8m + 16m + 32m = 63,5m
         private const val RUN_ATTEMPTS_FOR_63MIN_TOTAL = 7